+++ tools/cookbook.py
@@ -165,75 +165,3 @@ class MockServerTestCase(unittest.TestCase):
         '''Mock server_close().'''
         server.server_close()
         server.socket.close.assert_called_once_with()
-
-
-@unittest.skipIf(sys.version_info < (3, 3), 'unittest.mock since Python 3.3')
-@mock.patch('socketserver.BaseRequestHandler.handle', autospec=True)
-@mock.patch('socket.socket', autospec=True)
-class MockTCPServerTestCase(MockServerTestCase):
-
-    def mock_get_request(self, server_socket):
-        '''Override for TCP server.'''
-        server_socket.accept.return_value = (self.request, self.client_address)
-
-    def mock_server_activate(self, server):
-        '''Override for TCP server.'''
-        server.socket.listen.assert_called_once_with(server.request_queue_size)
-
-    def test_server_succ(self, MockSocket, mock_handle):
-        server = self.mock_server(MockSocket, mock_handle,
-                                  socketserver.TCPServer)
-        self.mock_handle_request(server)
-        self.mock_server_close(server)
-
-    def test_server_timeout(self, MockSocket, mock_handle):
-        timeout = 0.5  # in seconds
-        server = self.mock_server(MockSocket, mock_handle,
-                                  socketserver.TCPServer, timeout=timeout)
-        self.mock_handle_request(server, timeout=timeout)
-        self.mock_server_close(server)
-
-    def test_server_error(self, MockSocket, mock_handle):
-        error = KeyError
-        server = self.mock_server(MockSocket, mock_handle,
-                                  socketserver.TCPServer, mock_error=error)
-        self.mock_handle_request(server, error=True)
-        self.mock_server_close(server)
-
-
-@unittest.skipIf(sys.version_info < (3, 3), 'unittest.mock since Python 3.3')
-@mock.patch('socketserver.BaseRequestHandler.handle', autospec=True)
-@mock.patch('socket.socket', autospec=True)
-class MockUDPServerTestCase(MockServerTestCase):
-
-    def mock_get_request(self, server_socket):
-        '''Mock get_request() for UDP server.'''
-        data = 'mock data'
-        server_socket.recvfrom.return_value = (data,
-                                               self.client_address)
-        self.request = (data, server_socket)
-
-    def test_server_succ(self, MockSocket, mock_handle):
-        server = self.mock_server(MockSocket, mock_handle,
-                                  socketserver.UDPServer)
-        self.assertEqual(server.socket_type, socket.SOCK_DGRAM)
-        self.mock_handle_request(server)
-        self.mock_server_close(server)
-
-    def test_server_timeout(self, MockSocket, mock_handle):
-        timeout = 0.5  # in seconds
-        server = self.mock_server(MockSocket, mock_handle,
-                                  socketserver.UDPServer, timeout=timeout)
-        self.mock_handle_request(server, timeout=timeout)
-        self.mock_server_close(server)
-
-    def test_server_error(self, MockSocket, mock_handle):
-        error = KeyError
-        server = self.mock_server(MockSocket, mock_handle,
-                                  socketserver.UDPServer, mock_error=error)
-        self.mock_handle_request(server, error=True)
-        self.mock_server_close(server)
-
-
-if __name__ == '__main__':
-    unittest.main(verbosity=2, catchbreak=True)
